# docs/swagger.yaml
definitions:
...
  wn_internal_domain_dto.ExportInfo:
    properties:
      createdAt:
        type: string
      layouts:
        items:
          $ref: '#/definitions/wn_internal_domain_dto.Layout'
        type: array
      notes:
        additionalProperties:
          items:
            $ref: '#/definitions/wn_internal_domain_dto.Note'
          type: array
        type: object
      permissions:
        items:
          $ref: '#/definitions/wn_internal_domain_dto.Permission'
        type: array
      userId:
        type: string
    type: object
  wn_internal_domain_dto.ExportInfoRequest:
    properties:
      userId:
        type: string
    type: object
...
  wn_internal_domain_dto.ImportInfoRequest:
    properties:
      info:
        $ref: '#/definitions/wn_internal_domain_dto.ExportInfo'
    type: object
  wn_internal_domain_dto.Layout:
    properties:
      color:
        type: string
      id:
        type: string
      isMain:
        type: boolean
      ownerId:
        type: string
      permission:
        $ref: '#/definitions/wn_internal_domain_dto.Permission'
      title:
        type: string
    type: object
  wn_internal_domain_dto.LayoutIdRequest:
    properties:
      layoutId:
        type: string
    type: object
...
  wn_internal_domain_dto.NewLayoutRequest:
    properties:
      color:
        type: string
      title:
        type: string
    required:
    - color
    - title
    type: object
  wn_internal_domain_dto.Note:
    properties:
      draft:
        type: string
      haveAccess:
        items:
          type: string
        type: array
      id:
        type: string
      layoutId:
        type: string
      linkedWithIn:
        items:
          type: string
        type: array
      linkedWithOut:
        items:
          type: string
        type: array
      ownerId:
        type: string
      payload:
        type: string
      position:
        $ref: '#/definitions/wn_internal_domain_dto.Position'
      title:
        type: string
    type: object
  wn_internal_domain_dto.NoteId:
    properties:
      noteId:
        type: string
    required:
    - noteId
    type: object
...
  wn_internal_domain_dto.Permission:
    properties:
      canEdit:
        type: boolean
      canRead:
        type: boolean
      canWrite:
        type: boolean
      fromUserId:
        type: string
      id:
        type: string
      targetId:
        type: string
      toUserId:
        type: string
    type: object
...
  wn_internal_domain_dto.Position:
    properties:
      xPos:
        type: number
      yPos:
        type: number
    type: object
...
  wn_internal_domain_dto.UpdateLayout:
    properties:
      color:
        type: string
      layoutId:
        type: string
      title:
        type: string
    required:
    - layoutId
    type: object
...
  wn_pkg_response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/wn_pkg_response.meta'
      pagination:
        $ref: '#/definitions/wn_pkg_response.pagination'
    type: object
  wn_pkg_response.meta:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      requestId:
        type: string
    type: object
  wn_pkg_response.pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      perPage:
        type: integer
    type: object
info:
  contact: {}
  description: This is walrus notes api service.
  title: WALRUS NOTES API
  version: "1.0"
paths:
...
  /wn/api/v1/layout/create:
    post:
      description: Создать новый layout
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.NewLayoutRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.NoteId'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: create_layout
      tags:
      - layouts
  /wn/api/v1/layout/delete:
    post:
      description: удалить layout
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LayoutIdRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, cant_delete_main_layout,
            permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: delete_layout
      tags:
      - layouts
  /wn/api/v1/layout/export:
    get:
      description: Экспортировать лейауты, заметки, позиции и связи
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.ExportInfoRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.ExportInfo'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: export
      tags:
      - backup
  /wn/api/v1/layout/import:
    post:
      description: импортировать
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.ImportInfoRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: import
      tags:
      - backup
  /wn/api/v1/layout/my:
    get:
      description: Получить все layout-ы, к которым имеет доступ пользователь
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.NoteId'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/wn_internal_domain_dto.Layout'
                  type: array
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: get_my_layouts
      tags:
      - layouts
  /wn/api/v1/layout/update:
    post:
      description: обновить информацию о layout
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.UpdateLayout'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: update layout
      tags:
      - layouts
...
swagger: "2.0"